+++ resources/views/admin/pembina/index.blade.php
@@ -3,37 +3,24 @@
 @section('pembina')
 
 		<h5>Daftar Pembina</h5>
-		<nav>
-		    <div class="nav-wrapper">
-		      <div class="col s12">
-		        <a href="#!" class="breadcrumb">First</a>
-		        <a href="#!" class="breadcrumb">Second</a>
-		        <a href="#!" class="breadcrumb">Third</a>
-		      </div>
-		    </div>
-		</nav>
+		<hr>
+
  	 {{-- Menu --}}
 		<div class="row">
-			<a href="{{ route('pembina.tambah') }}" style="margin-top: 20px;" class="waves-effect wave-light btn right"><i class="fa fa-plus"></i> Tambah</a>
+			<ul class="tombol">
+				@if(Auth::user()->level == 'admin')
+					<li>
+						<a href="{{ route('pembina.tambah') }}" class="waves-effect waveslight btn blue accent-2"><i class="fa fa-plus"></i> Tambah</a>
+					</li>
+					<li><a href="" class="waves-light waves-effect btn amber accent-4"><i class="fa fa-download"></i> Import</a></li>
+				@endif
+				<li><a href="" class="waves-light waves-effect btn grey lighten-1"><i class="fa fa-print"></i> Print</a></li>
+			</ul>
 		</div>	
+
 		<div class="row">
 			<div class="col m12">
 				@include('templates.alert')
-				{{-- Search --}}
-				{{-- <div class="row">
-					<div class="col m10 z-depth-1" style="border: 1px solid #f3f3f3">
-						<form action="{{ route('pembina.search') }}" method="get">
-							<div class="input-field">
-								<input id="search" type="text" name="search" placeholder="Search Data" required>
-							</div>
-					</div>		
-						<div class="row">
-							<div class="col m2">
-								<button type="submit" class="btn btn-large" style="margin-top:10px;"><i class="fa fa-search"></i></button>
-							</div>
-						</div>
-						</form>
-				</div> --}}
 
 				{{-- Table --}}
 				<table id="TableId" class="mdl-data-table" cellspacing="0" width="100%">
@@ -44,7 +31,9 @@
 						<th>Jenis Kelamin</th>
 						<th>Alamat</th>
 						<th>Opsi</th>
-						<th>Hapus</th>
+						@if(Auth::user()->level == 'admin')
+						 <th>Hapus</th>
+						@endif
 					</tr>
 				</thead>
 			
@@ -56,27 +45,27 @@
 							<td>{{ $datas->jenis_kelamin }}</td>
 							<td>{{ $datas->alamat }}</td>
 							<td>
-								<a href="{{ route('pembina.lihat', $datas->id) }}" class="waves-light waves-effect btn-floating"><i class="fa fa-eye"></i></a>
-								<a href="{{ route('pembina.edit', $datas->id) }}" class="waves-effect waves-light btn-floating"><i class="fa fa-pencil"></i></a>
+								<a href="{{ route('pembina.lihat', $datas->id) }}" class="waves-light waves-effect btn-floating green"><i class="fa fa-eye"></i></a>
+								@if(Auth::user()->username == 'admin')
+									<a href="{{ route('pembina.edit', $datas->id) }}" class="waves-effect waves-light btn-floating amber"><i class="fa fa-pencil"></i></a>
+								@endif
 							</td>
-							<td>
-								<form action="{{ route('pembina.hapus', $datas->id) }}" method="delete">
-									<input type="hidden" name="_token" value="{{ csrf_token() }}">
+							@if(Auth::user()->username == 'admin')
+								<td>
+									<form action="{{ route('pembina.hapus', $datas->id) }}" method="delete">
+										<input type="hidden" name="_token" value="{{ csrf_token() }}">
 
-									<button type="submit" name="hapus" class="waves-effect waves-light btn-floating"><i class="fa fa-trash"></i></button>
-								</form>
-							</td>
+										<button type="submit" name="hapus" class="waves-effect waves-light btn-floating red"><i class="fa fa-trash"></i></button>
+									</form>
+								</td>
+							@endif
 						</tr>
 					@endforeach
 				</tbody>
 			</table>
 			</div>
 		</div>
-		{{-- <div class="row">
-			<div class="col m12 offset-m2">
-				{{ $data->render() }}
-			</div>
-		</div> --}}
+		
 	</div> {{-- End Menu --}}
